Add files via upload
Finally figured out how to make a login method with Node. When you click on the New User button, it will let you submit your info and it stores them to an array. Then once you try to login, it will verify your input with that is stored in the array.